Use equality not containsAll to compare list of annotation values
Previously, the `equals` methods on `AnnotationArrayAttributeValue`
implementations would use the `containsAll()` to compare the list of
values. That has a couple of problems.
1. It ignores the order of values in the array. While that may be ok
   for some annotations it is not true for all of them.
2. It ignores differences between annotations where both have multiple
   values, are not equal but where one is a subset of another.

This change adds additional tests to highlight the second issue. It
adds tests to cover when the first item is a subset of the other and
vice versa. When using `containsAll` the test fails on one of the new
fields.

This change then switches to use equality and corrects the existing
test that treats `"2", "40"` the same as `"40", "2"`.

Bug: 291562733
Test: ./gradlew
Change-Id: I40d5b2008965f7581f6e1ca35e826894bb9de463