Add a way to force non-staged APEX update for any APEX
This is a development only feature that only works on debuggable builds.
It can be used to speed up development workflow for teams that have
their code packaged in an APEX.

Example of how this feature can be used:
1. Iterate on code in an APEX
2. Build APEX
3. adb install --force-non-staged out/dist/your.apex
4. Restart the processes that depend on this APEX (e.g. adb shell stop
   && adb shell start).
5. ???
6. Profit

Behind the scenes the force non-staged APEX update is implemented by
unmounting the /apex/ mount point with MNT_DETACH flag.

Bug: 290750901
Test: atest ApexTestCases
Test: build com.android.virt APEX
Test: adb install --force-non-staged out/dist/com.android.virt.apex
Merged-In: I20db9a9878f7b280599c124e0f516adc0430ed20
Change-Id: I20db9a9878f7b280599c124e0f516adc0430ed20