Make DefaultAnnotationValue sub-classes more general
Previously, the `DefaultAnnotationValue` sub-classes were limited in
how they could be used. Their constructors expected the `valueSource`
to be provided and would parse it in a specific way. This change
extracts the parsing into the `DefaultAnnotionValue.create()` method
and changes the sub-classes to take getters for both the source and
value. The `value` and `valueSource` are initialized lazily from their
respective getters.

Taking getters will allow the sub-classes to be used in a variety of
situations. e.g. A follow up change will open the classes up to be
extended for `PsiAnnotationValue` and `UAnnotationValue` both of which
defer getting the source and value, albeit inconsistently depending on
whether it is a single value or an array of values.

Bug: 291562733
Test: ./gradlew
Change-Id: I5c9f31adc1f5da19679e328e3ae41de8590aa2e2