Pull originalName/qualifiedName to DefaultAnnotationItem
Previously, each sub-class of `DefaultAnnotationItem` provided its own
implementation of the `originalName` and `qualifiedName` properties.
This change pulls those implementations up into
`DefaultAnnotationItem`.

The `mapName` parameter is needed because in some cases the
`TextBackedAnnotationItem` does not map names to avoid converting
recently nullability annotations back into normal nullability
annotations.

The second constructor is needed because the initializer of the
`qualifiedName` property uses the `codebase`. That works fine when the
`codebase` parameter is accessed but causes a warning when the
`codebase` property is accessed. That is because the `codebase`
property is overridden by sub-classes and so may not be initialized at
the time it is used in the initializer.

Moving the property definition of `codebase` out of the primary
constructor avoids that (as then the initializer is using the
`codebase` parameter) but then it reports a warning saying that the
`codebase` property could be defined in the primary constructor.

Making the primary constructor private and then adding a secondary
constructor that computes `qualifiedName` using the `codebase`
parameter avoids those issues.

Bug: 291562733
Test: ./gradlew
Change-Id: Ia75f84199f946b047b6fb245b75787c6ef7a140a