utils: eth_mac_addr: Refactor ethernet mac addr script
Refactored the scripts which we use on DB845c to set
Ethernet MAC address.

Updated the .sh script to fall back to parsing bootconfig
file to lookup androidboot.serialno bootarg, if it is not
found in /proc/cmdline (boot image header v4 for example).

Fixed a typo s/db45c/db845c in .rc script which was blocking
this service to run.

Renamed the script from eth_mac_addr to set_ethaddr to
follow the nomenclature we have used everywhere else and
moved them to common shared/utils/ directory.

And updated the sepolicies accordingly.

Change-Id: Ia25e464b7c13934ffd8ab66a6aa2d892d7cd5a7b
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>