Autoupdate of 2 submodules
*external/adt-infra updated to f5e596b1b143f398c6db8472177f6a8319bd922b
Update boot tests with notification

This adds an integration tests that validates we now deliver boot
notifications over gRPC

It also fixes as subtle bug in the wait_until function where we secretly
assumed that if a predicate returns true, it will always return
true.

This was due to the double invocation during exit, which could make it
possible for this to return False, even though we didn't timeout yet.

Change-Id: I4bcd927bc86430077d94f888bc8fbba734b3a49b

*external/qemu updated to 409bec988ab423e07b4385e8be15d0a165185bb7
Merge changes from topic "grpc-notifications" into emu-master-dev

* changes:
  gRPC: Add boot notification events
  gRPC: Make notifications async