[scudo] Clean up tests.
Modify the tests so that all clang warnings can be turned up to high.

Fix all places flagged by -Wconversion.

Fix a few unused variables not marked with UNUSED.

For the memtag testing, only compile some tests for 64 bit since
compiling them on 32 bit leads to warnings/errors. All of the tests
are already skipped on 32 bit OSes, so this will not affect any
real tests.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D155749

GitOrigin-RevId: af41f79f40f454374f63747d4b53afde296189fe
Change-Id: I3004f53ce495f8b0cc9e8838a277ec9bf6647379