Add a debugging switch for trunk stable configuration
It is important that clients and services both behave like the last
frozen version of the interface, so both sides have protections.
In order to test each side seperately, this switch is introduced.
READ means only the parcel reads and the onTransact in the Stub/Bn*
classes contain the protections.
WRITE means only the parcel writes and sending the transactions
contain the protections.

Test: atest aidl_integration_test with all three options
Bug: 290265210

Change-Id: I4b097d5bcea47af3ff162e751a64d3e5eb6c6454