Force compilation job to run
The run command says "For historical reasons, some constraints, such
as battery, are ignored when this command is called", and the -f
option makes it "run the job even if technical constraints such as
connectivity are not currently met."

We've got evidence that in some environments the job is not getting
scheduled, but the tests needs it to run. So it makes sense to ignore
all the job constraints.

Bug: 270528861
Test: atest CompOsDenialHostTest
Change-Id: I6c827f0a90c54df98b0ff4756016009bc9e10bd8