Split AnnotationManager.mapName() into separate input/output methods
Previously, `mapName()` was used for a couple of separate purposes:
1. Normalizing annotation names on input.
2. Filtering annotations from the input, e.g. because they are
   unsupported or not needed or explicitly excluded.
3. Customize the output name depending on the target file.

That makes the method difficult to understand and inefficient because
it is doing more work than needed. So, this change starts by replacing
the single `mapName()` method with separate methods for the input and
the output. At the moment they both delegate to the existing
`mapName()` but following changes will split the functionality out into
the separate methods.

Bug: 292484829
Test: ./gradlew
Change-Id: I44d744e7ebca13488e871625c969bc3f25ab23cb