Move RecentlyNullable/NonNull rewriting to normalizeOutputName
Previously, the `RecentlyNullable/NonNull` annotations were mapped to
AndroidX `Nullable/NonNull` annotations on input and then on output the
`Psi/UAnnotationItem`s would pass the `originalName` into the
`mapName()` method, along with `target = SDK_STUBS_FILE` to cause the
`RecentlyNullable/NonNull` annotations to be written to the stubs file.

This change keeps the `RecentlyNullable/NonNull` annotations in the
model, and only maps to AndroidX `Nullable/NonNull` on output, allowing
the `Psi/UAnnotationItem`s to pass `qualifiedName`. Keeping the
recently annotations in the model does not change the behavior as they
are still treated as nullable/non-null just as if they had been mapped
to AndroidX nullability annotations.

Bug: 292484829
Test: ./gradlew
Change-Id: Ied5275be697e3b795f81ddb937cefb941f11c700