Always normalize annotation names on input
Previously, a number of places that created annotations disabled the
mapping of input names as they replaced `RecentlyNullable/NonNull`
annotations with AndroidX `Nullable/NonNull` annotations which was not
what that code needed. So, they passed `mapName = false` to prevent
that from happening and preserve the `Recently` nullability annotations
in the model. However, a preceding change made that the default
behavior so as there is no longer a need to pass in `mapName = false`
this change removed the `mapName` parameter.

The `PsiClassItem` passed in `mapName = false` because the `mapName()`
had no effect on the `java.lang.annotation.Retention` annotation so it
made no sense to do the extra work for no reason. However, there is no
point in maintaining the `mapName` parameter simply for that one case.

Bug: 292484829
Test: ./gradlew
Change-Id: I12d26037edebc2312b64ca80a185c8178ad93e7d