Stop calling mapName() from normalizeOutputName()
Previously, `mapName()` was called from `normalizeOutputName()` to
preserve the existing behavior which functionality was moved out to
the `normalizeOutputName()`. This change removes that call and inlines
`mapName()` into `normalizeInputName()` as that is the only other
caller.

Removing the call from `normalizeOutputName()` to `mapName()` caused
a failure in the `NormalizeOutputAnnotationTest` as the names of nested
annotations were no longer normalized. So, this change made an explicit
call to `normalizeInputName()` for the names of the nested annotations.

Note, only the change to `PsiAnnotationItem` was needed to fix the
breaking test. The change to `UAnnotationItem` was made for
completeness as it was not possible to create a similar test that used
the `UAnnotationItem` class.

Bug: 292484829
Test: ./gradlew
Change-Id: I6ba8fdfe4a092022bcefd3b7d10bd6899ae5e3a4