Cache show annotation status
Extends the `AnnotationInfo` caching mechanism to support caching
whether an annotation is a show annotation or not. The status is stored
in a new `AnnotationItem.show` property.

That required the following enhancements to the caching mechanism:
1. Annotations that can be matched by `AnnotationFilter`s must use
   the whole annotation source (including attributes) as filters can
   match against annotation attributes as well as the `qualifiedName`.
2. A subclass of `AnnotationInfo` was needed which could compute the
   value of the `show` property lazily based on the `config`.

Existing uses of the `options.showAnnotations.matches(annotationItem)`
function have been replaced with `annotationItem.isShowAnnotation()`.

Bug: 292484829
Test: ./gradlew
Change-Id: I2db0ab25c84ba07afdd9763d96cda3e5b58d0cb9