Add new golden output for trunk stable behavior
The 'next' configuration has the build flag set to True while the other
configurations are set to 'False'. This flag determines the arguments to
the aidl compiler and the generated code. In order for the golden_test
check command to work for both targets in the same code base, we now
have two different sets of golden files.

Note: this is temporary, as we plan to call into a flagging library in
the generated code once they are available in C++ and Rust. Once that
happens, there will be no difference between the generated code in
'next' and the other configurations so this workaround can be removed.

Test: set RELEASE_AIDL_USE_UNFROZEN=false
Test: set RELEASE_AIDL_USE_UNFROZEN=true
Test: atest aidl-golden-test-build-hook
Test: AIDL_USE_UNFROZEN=<true/false> golden_test.sh update

Bug: 290265210
Change-Id: Ia09fc3ccc17fd91d98c0e76f04fe72bb21ef5237