Cache hide/hideMeta in AnnotationInfo
Previously, whenever anything needed to know whether an annotation was
a hide annotation it would apply the annotation filters to the
annotation. This change caches that information in the `AnnotationInfo`
instance associated with the `AnnotationItem`.

This adds a mechanism to detect and break out of cycles caused by an
annotation being applied to itself, or to another annotation that is
then applied to itself. That is a potential problem when checking an
annotation involves resolving it to its definition and the looking at
its annotation. See `isComputingSuppressCompatibilityMetaAnnotations`
for an example.

The `hideMeta` should not cause that problem because it only checks one
level but it could interact with other code to cause a problem.

Bug: 292484829
Test: ./gradlew
Change-Id: I736e7ba90d3a6056a9642968936dd53ca0cc4058