[scudo] seperate cache retrieval logic
Split cache::retrieve() into separate functions. One that retrieves
the cached block and another that sets the header and MTE environment.
These were split so that the retrieve function could be more easily
changed in the future and so that the retrieve function had the sole
purpose of retrieving a CachedBlock.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D155660

GitOrigin-RevId: 4c6b8bb87b3452d0bcef83cd0ea712d8426603b8
Change-Id: Ic3f18065acd0d20f3361d4a802afb2540c62e92a