ANDROID: run cargo_embargo after merge plus fixes
Fixes: add wire_format_derive_test_src_lib to cargo_embargo.json
module_blocklist because its dependencies don't have any android
variants.
Remove the riscv64 Android.bp.pach because it failed to apply and was
not needed.
Remove the seccomp assignment of tgkill in jail_warden.policy because it
was causing compilation to fail due to being assigned in two places
(device_common.policy being the other place).
Remove the perfetto directory and remove the .gitmodules file that the
directory/module was added to.

Test: m
Bug: 292270337
Change-Id: I6d1a3d7f19626ac2cd74c99e3f79f48ea70f7235