Cache suppressCompatibility in AnnotationInfo
Previously, whenever anything needed to know whether an annotation was
a suppress compability annotation it would apply the annotation filters
to the annotation. This change caches that information in the
`AnnotationInfo` instance associated with the `AnnotationItem`.

This makes use of the new mechanism to detect cycles added in the
previous change and removes the existing mechanism.

Bug: 292484829
Test: ./gradlew
Change-Id: I654bca9c5f20590c614200eac733826d8cfbe9e6