Enable R8 compat for SdkSandbox app
This explicitly preserves existing behavior, and should be a no-op
with respect to current compilation. As R8 will soon be the global
default, an explicit setting is required to avoid issues with
optimizations in R8 full mode use of reflection with SandboxedSdkHolder
These issues have been resolved elsewhere but not yet upstreamed.

Bug: 293177283
Test: atest SdkSandboxManagerTests
Change-Id: I481c74db37e273f0feaa13d3ac94ca0c3f9d3bbb