Add --remote_ssk_key
If --remote_virtual_devices is specified. the CLI tests the SSH connection from the ATS host to the remote host with the private key. Connection failure is shown as a warning message and does not stop the CLI. The user can fix it after the container starts.

The CLI copies the key to a specific path in the container. The device action for remote virtual devices will refer to the path.

PiperOrigin-RevId: 551099109
Change-Id: Ic2095b551e47a1daaccb35249963cc755fda5b0f