Delegate hide annotation checks to AnnotationManager
Previously, `ModifierList` implemented those checks itself which meant
that it had to depend on `options`. This change moves the checks into
the `AnnotationManager` implementation removing some of those
dependencies.

As part of this change the `hasHideMetaAnnotation()` check becomes an
implementation detail of the `DefaultAnnotationManager`.

This is part of a series of changes that will completely remove the
dependency on `Options` and `options` from `ModifierList`.

Bug: 292484829
Test: ./gradlew
Change-Id: Icd9297c2e28d6e4b8e0f640779a8de0c0fa8b6d1