Make MutableAnnotationFilter immutable
Previously, it was not clear whether the `add()` method was ever called
after calling the other methods, specifically `matchesAnnotationName()`
which cached the result of calling `getIncludedAnnotationNames()`.

This change makes that clearer by moving the `add()` method out into a
separate `AnnotationFilterBuilder` class and renaming
`MutableAnnotationFilter` to `ImmutableAnnotationFilter`. It also
simplifies the caching by using `lazy`.

Bug: 292484829
Test: ./gradlew
Change-Id: I3641e8aa9b667728c68a882caf613443297f862d