Prevent Codebase mutation through interface
Previously, `Codebase` had a few `var` properties which made it
mutable. This change makes those properties `val` and moves any code
that modifies them down into the `DefaultCodebase`. It also dedups
the overrides of those properties across the various subclasses of
`DefaultCodebase`.

This does not make the implementations of `Codebase` immutable it
just prevents mutation through the `Codebase` implementation.

Bug: 293345573
Test: ./gradlew
Change-Id: I2e37fbc888aaf1fd0a94a6eb598492e2c34c6dd0