Autoupdate of 2 submodules
*build/soong updated to f97a42e5f2a5a9aea863660487d39c6b672a2fa4
Merge "APEXes contain VNDK libraries when VNDK is deprecated" into main
*system/linkerconfig updated to a5322f57ee2dd0ec7fd2d0089db7d9a6df2295e0
Vendor APEX do not link to VNDK when it is deprecated

From prior work, VNDK namespace was still active from Vendor APEXes
because some of them relied on VNDK libraries by linking. However, this
is not available when VNDK is deprecated. This change removes link from
Vendor APEXes to the VNDK when the VNDK is deprecated.

Bug: 290318998
Test: Checked from AOSP CF x86_64 Phone build with KEEP_VNDK=false that
VNDK namespace has been removed with vendor APEXes

Change-Id: I3bcb9aee6a16e25cdaa428572f34d15c18bd7f14