Split Reporter into separate interface and implementation
Previously, the `Reporter` implementation class was used through
metalava code which caused that code to transitively depend on lots of
implementation specific classes. This change splits the class into
an interface (still called `Reporter` to avoid churn elsewhere in the
code) and a new `DefaultReporter` implementation class.

Splitting requires some unfortunate downcasting to access
implementation details of the `DefaultReporter` because the callers and
the implementation both directly access the same `Options` properties.

Bug: 293458631
Test: ./gradlew
Change-Id: I78aa34f953cc381a2cee9d608302325787235908