Encapsulate the code for fully qualifying Javadoc into a class
Previously, the code for fully qualifying Javadoc was just a group of
functions. This change encapsulates that in a new `DocQualifier` class.
That allows them to shared a `reporter` property which is passed into
the constructor, thereby avoiding the need to have to modify all the
methods if the `reporter` was passed as a parameter.

Bug: 293429095
Test: ./gradlew
Change-Id: I3850710d0c8282f3bf18b02e2aec60f6e3928935