Fix incorrect escaping of string literals
There is the potential that this could cause churn in signature files
and stubs files as the affected code is used when generating them.
However, it is likely that there are few differences as the output
would be incorrect and should have been caught by reviews of the
signature files. The value output in the stubs files would also be
incorrect so likely to cause App bugs too.

If stubs are affected then fixing this could cause App compatibility
issues for applications that rely on the wrong value.

The resulting output is not quite the same as how IntelliJ would
escape the values as `'` is escaped unnecessarily which is visible in
IntelliJ but it is still correct and changing that could also cause
churn in signature files.

Bug: 293537075
Test: ./gradlew
Change-Id: I91fc25c2cf39cc16098fe104ee3651780eb47f4a