Pre-compute method infos before flushing trace entries to file
Computing method infos requires mutator lock. We used to compute method
infos when flushing the trace entries to file. This meant we were
holding mutator lock while we flush entries to file. This changes it so
that we pre-compute method infos. That way we only need mutator lock for
a short time and don't do File I/O while holding the lock.

Bug: 259258187
Test: art/test.py

Change-Id: Idd7e3be97d5a35f7323c019ea953d4aaf7acf9b3