Enable HiddenAbstractMethod for showSingleAnnotations
Previously, the method was disabled for `showSingleAnnotations` as
there were issues with the `@IntraCoreApi` and possibly
`@CorePlatformApi` surfaces which used the `--show-single-annotation`
option. The last such issue is fixed in https://r.android.com/2685827
so it is safe to enable this check fully.

As that was the last use of the `Options.showSingleAnnotations`
property from outside the `Options` class this marks the property as
`private`.

Bug: 292484829
Test: ./gradlew
Change-Id: I96df98551cb5e8c5d831fc3f9d9595f52cefb349