Improve test coverage for adding final modifier to class and method
Splits the single `Add final` into separate tests for class, method and
method parameter. The class and method tests are themselves split into
two, one for classes that can be extended and one for classes that
could not be extended.

The tests make the `AddedFinalUninstantiable` issue an error so that it
will be reported as in issue in the test as normally it is hidden. That
ensures that it is being reported properly and not just being ignored.

Bug: 294300564
Test: ./gradlew
Change-Id: I657bbcf59155dadb3dee95797d5200c5c53315d3