Use ClassItem.getRetention() instead of hasSourceAnnotation
The `ExtractAnnotations.hasSourceAnnotation(...)` methods were
duplicating functionality for the `psi` model that is provided
generically via the `ClassItem.getRetention()` method. This change
replaces that duplicate functionality using `getRetention()`.

One possible change in behavior is that previous
`hasSourceAnnotation(...)` would always return `false` for
`TextClassItem` but now it will return the correct retention.
However, if it did return `false` that would be treated as an error so
it is unlikely that this code was ever called on a `TextClassItem`.

Bug: 293429095
Test: ./gradlew
Change-Id: I3c06e06da625892505b11ca490f71c080e5bcafb