Test: Catch all Nat64PrefixUpdates in PrefixDiscoveryBypassTls
The test PrefixDiscoveryBypassTls expects DnsResolver to trigger
onNat64PrefixEvent 4 times (2 for prefix-added events; 2 for
prefix-removed events) , but it doesn't catch all of the event
before it ends. This will cause the next test SetAndClearNat64Prefix
to fail if the last onNat64PrefixEvent is received after
SetAndClearNat64Prefix runs.

This change improves PrefixDiscoveryBypassTls to ensure that all
onNat64PrefixEvents are caught before the test ends.

Bug: 293833962
Test: ran PrefixDiscoveryBypassTls and SetAndClearNat64Prefix together
      300 times
Change-Id: I09de58b957679602fae2c9be1caa1e7f0fa29c21