Create new testsuite package
The `CommonTypeItemTest` was supposed to be the beginning of a set of
common tests for implementations of the `model` to ensure consistent
behavior, where applicable. Unfortunately, since that was added things
have changed a little.

The main issue is that the `MethodItem.internalDesc()` method has been
extracted into an extension function in `AddApisFromCodebase.kt` so is
no longer part of the model.

This change adds a separate test for that method and then renames the
`CommonTypeItemTest` to `CommonMethodItemTest` (which would have been
more appropriate anyway) and modifies it so it does not depend on
`MethodItem.internalDesc()`. It also moves it into a more appropriate
package.

Bug: 294556178
Test: ./gradlew
Change-Id: I5295960fdd907de60b531f5640069f507dd7b884