Add PsiEnvironmentManager to keep track of UastEnvironments
Previously, that was done through a global variable `uastEnvironments`
and some top level functions. This change abstracts the global variable
and functions within a `PsiEnvironmentManager` which implements
`Closeable` to allow it to be used with `Closeable.use`.

Note, the existing code is not modified and the `PsiEnvironmentManager`
delegates to it to reduce the size of this change. This change focuses
on integrating the `PsiEnvironmentManager` into the code. A follow up
change will cleanup the global variable and associated functions.

Bug: 294725097
Test: ./gradlew
Change-Id: If3cb8cbfe9353910df763f257b699f41b1d8f5c6