Extract initialization of subcommands from MetalavaCommand
Previously, the `MetalavaCommand` would initialize its `subcommands`
itself which meant that it depended on all the subcommand classes and
they in turn depend (albeit indirectly through the context such as for
help formatting and the console) on the `MetalavaCommand`. That would
tightly couple them making it hard to modularize.

This change pulls the initialization of the subcommands out into the
`Driver.kt` removing the dependency from `MetalavaCommand` onto those
subcommands.

Bug: 294815714
Test: ./gradlew
Change-Id: I578e67a435f0acf5e2582af43d5ad8d2e044930b