Simplify adding test classes to the metalava-model-testsuite
Previously, the sole testsuite test class was an abstract class that
required the model implementation providers to extend the class and
implement the `createCodebaseAndRun` method. That was not particularly
scaleable as it would require a new class and an implementing method
for each combination of test suite class and model implementation.

So, every time a test class was added to the suite it would require an
additional two test classes added for text and psi implementations. Ir
was also error prone as if a class was not added then the test would be
ignored.

This change removes the need for the model implementation to extend the
class by using a service provider interface structure. The
`createCodebaseAndRun` method is moved into the `ModelSuiteRunner`
interface. Each test is a JUnit `Parameterized` test that takes an
of `ModelSuiteRunner` which it then uses to run the tests. The list of
`ModelSuiteRunner` instances to pass as parameters is found by using
`ServiceLoader` to load them.

Each model implementation will provide its own implementation of the
`ModelSuiteRunner` just like it previously provided an overriding
method. Once that is done each model implementation will run all the
tests and any new ones which are added.

One slight complication is that gradle will not run tests that are
provided by a dependency. This works around that by creating a special
`metalava-model-provider-plugin` which must be applied by a project
that provides an implementation of the model and which will unpack the
`metalava-model-testsuite` to make it look as though they were built
by the project.

An attempt was made to use `@AutoService` to handle the generation of
the `META-INF/services/...` file but unfortunately, the annotation
processor could not be made to work so they were just created by hand.

Bug: 294556178
Test: ./gradlew
Change-Id: I28b9f0a0db5076461bd6a8ac6a392a3b08ee90fb