Add an option to AppLaunchTest to enable blank screen detection with threshold.
The option allows to specify the threshold for the percentage of blank screen which will fail the test.
We use the BlankScreenDetectorWithSameColorRectangle, wich computes the percentage of the screen occupied by the largest same-color rectangle.
When the blank screen area is bigger than the threshold, the test fails and the screenshot with the rectangle outline is added to the tests artifacts.

Test: atest  csuite-harness-tests
Bug: 292125742
Change-Id: I1ef355db53517739e608249a6144ff9d081a98d8