Extract findKotlinStdlibPaths from findKotlinStdlibPathArgs
The `findKotlinStdlibPathArgs` function returns a String array that can
be passed on the command line for parsing by `Options.parse()` which
means it is not suitable for use by tests that cannot depend on
`Options`. This change extracts `findKotlinStdlibPaths` (which just
returns the paths as a `List<File>`) from that function and moves it
into `...metalava.testing`.

Bug: 294725097
Test: ./gradlew
Change-Id: I637c6738f813ca430f0ddc39c9c50c33f526fb74