Move ...model.psi into metalava-model-psi
Moves the main source for `...model.psi` into `metalava-model-psi` but
leaves the test code for the moment as it needs the `prebuilts/sdk/...`
artifacts but they are only usable from `:metalava:test`.

The main source is being moved without the tests because it will
prevent any invalid dependencies being added to the `psi` code while
the tests are being fixed which is being done in b/295182989.

Some issues were moved from the `lint-baseline.xml` file to the
`metalava-model-psi/lint-baseline.xml` file.

Bug: 287241408
Test: ./gradlew
Change-Id: Ia850342d5b186bdd542acf775ac5cbdfe2fca1c1