Merge defaultConfiguration and configuration global properties
Previously, the `defaultConfiguration` was a `val` that was initialized
to an `IssueConfiguration` and `configuration` was a `var` that was
initialized to the value of `defaultConfiguration` but was never
changed. Some parts of the code used one and some the other which was
very confusing. Also, `configuration` is a pretty generic name.

This change merges them into a single `val issueConfiguration`
property.

Bug: 295013652
Test: ./gradlew
Change-Id: I04ddc16ac8639f0f3353de4dfa0471cefbae255a