Move global issueConfiguration to Options
This makes it easier to track which `issueConfiguration` is being used,
makes iteasier to identify code that needs to be cleaned up and makes
it easier to move the `issueConfiguration` creation to be driven by
Clikt.

In `CompatibilityCheck` it was cleaner to pass the `issueConfiguration`
in as a constructor parameter.

The `Options.setIssue/CategorySeverity` had to be pulled out of the
companion object and made instance methods as they are called while
executing `Options.parse` but at that point the global `options`
variable has not yet been updated to refer to the `Options` on which
`parse` is being run. So, if those methods attempted to access the
`IssueConfiguration` to update via `options.issueConfiguration` then
they would end up updating the `IssueConfiguration` for the wrong
`Options` instance and those updates would be discarded upon return
from `parse`.

Bug: 295013652
Test: ./gradlew
Change-Id: I3300f7d8cc1ca740bc2f6198150da5a7ee66120d