Add OptionGroup for signature related options
Previously, the `Options.usage()` method had the ability to group
options into sections with a title. Clikt also has the ability to do
that using named `OptionGroup`s`. This change adds an `OptionGroup` for
signature related options and moves a couple of the existing options
from the current group.

Adding the first named `OptionGroup` (the existing `OptionGroup`
classes have no name) revealed some strange behavior in Clikt where it
sorts any option with no group name after all the groups. That lead to
some weird formatting so it was necessary to force all options to have
a group name.

Bug: 295311787
Test: ./gradlew
Change-Id: Ib9dd1590450cd9f031fec58701292a4efc31a8f9