Remove IssueConfiguration.reset()
Previously, when the `IssueConfiguration` used by the reporters and
other parts of metalava was accessed via a global val it was necessary
to reset the state of that before running a test. However, now that it
is owned and created by `Options` that is no longer necessary as the
global `options` property is reset every test anyway.

Bug: 295013652
Test: ./gradlew
Change-Id: I08a190d3bf51fc8b797c67c18391baac440dae13