Make --output-kotlin-nulls=yes require --format=v3 or above
Previously, if `--output-kotlin-nulls=yes` was specified and the
`outputFormat` was less than `V3` then it would change `outputFormat`
to be `V3. Changing the `outputFormat` based on the setting of one of
the values that it configures on the `Options` makes it quite hard to
reason about. So, instead this change simply verifies that the
`outputFormat` is `V3` or above.

Bug: 295311787
Test: ./gradlew
Change-Id: Id6087be2927c86a13135ce454d2c8471d3a33dd7