Sort out the CommonOptions handling
Previously, the `CommonOption` was supposed to be passed to the
`Options` class so it could delegate its `quiet`, `verbose` and
`terminal` properties to it. Unfortunately, that did not work due to
some code accessing `Options` without parsing it through Clikt and so a
temporary fix was added to use a `defaultCommonOptions` instance which
was initialized through a `FakeCommand`.

Since then a lot of changes have been made which means it is now
possible to fix this issue properly. There are a number of different
changes needed to make that happen:
1. Was the realization that there is no way to have `Options` and
   `CommonOptions` set their `quiet` and `verbose` properties in the
   same way (at least not at the moment). That is primarily because the
   `Options` properties are set early by `preprocessArgv`. Having them
   initialized separately, albeit from the same arguments makes it a
   little easier to handle.
2. A `verbosity` property was added to `Options` to make it more
   consistent with `CommonOptions`. The existing `quiet` and `verbose`
   are define in terms of that, just like in `CommonOptions`. This
   property is set by `preprocessArgv` which makes that a bit simpler.
3. `CommonOptions` needed to be tunneled through to `OptionsCommand`
   to pass into the `Options` constructor for use by the
   `Options.terminal` property.
4. The `Options.terminal` has to delegate to `CommonOptions.terminal`
   instead of just retrieving it as it will not have been initialized
   before being passed to `Options`.

Those changes allows the `defaultCommonOptions` to be removed.

Bug: 286023667
Test: ./gradlew
Change-Id: Ie300a988c44e4d6c216f99fb8864d11dc8a4d454