Remove testBoot[Disable|Enable]PKVM tests
The tests were to prove that the overhead of pKVM is minimal. Since pKVM
on Pixel is the default, we no longer need to prove ourselves to Pixel.

In addition, this test relied on a Pixel-specific `fastboot oem pkvm`
which is not guaranteed to exist in partner devices, so it has little
value to our partners. Eventually, the switch will disappear even in
Pixel.

Therefore, I am removing the tests.

Bug: 294503342
Test: build
Change-Id: Ie598ca8afa8034e24d12d7d04a06cdeffbff7760