Move ApiFile.parseHeader back into FileFormat
A previous change moved it out but it seems as though it would be best
if it was in this class alongside the code for creating the header that
is written into the file. This also makes a few other minor changes:
1. Pushes the buffering creation into the method to ensure consistency
   across the different callers.
2. Uses a `LineNumberReader` instead of a `BufferReader` to provide
   access to line numbers for better error reporting.

This is in preparation for further work to support a new signature
header format.

Bug: 295577788
Test: ./gradlew
Change-Id: I46181b620fba49d3b94f93b84e72826ee1daf019