Removes unnecessary stripBlankLines and trim parameters
The parameters on `DriverTest.check` were always true and passed
through to the corresponding parameters on `readFile` which were also
always true (apart from one minor exception below) and so they were
inlined.

The exception was that proguard file called `readFile` without
providing the parameters so they were `false`. However, they have no
effect on the proguard file contents so it did not matter anyway. There
was also a slight ordering issue with an assertion to check the
existing of the proguard file running after the code to read the file
so they were re-ordered.

Bug: 295013652
Test: ./gradlew
Change-Id: Ia4a0923d0d5b5ca1dac25fa03acd085d72ad6ccf