Make DriverTest format handling consistent with CLI processing
As a convenience `DriverTest.check` will automatically add a signature
header to a test that is checking the API output but does not provide
its own signature. That reduces duplication for those tests that are
agnostic about the signature version. Previously, the calculation of
the format in the test was different which meant format changes would
cause test failures for those tests. This change extracts that logic
for how to construct a `FileFormat` from various command line arguments
into a `FileFormat.applyOptionalCommandLineSuppliedOverrides` method
that is used from both the CLI and the test.

The new method encapsulates the special behavior of the override for
`overloadedMethodOrder` compared to `kotlinStyleNulles` which will make
it easier to test and modify in follow up changes.

Bug: 295577788
Test: ./gradlew
Change-Id: Id54732de1c98d9dbc1a39a0f0802d70ac6f59266