Fix tests that would be broken by adding a new FileFormat version
Adding a new version changes the value of `FileFormat.LATEST` which
would break any tests that did not specify the format to output but did
include the signature for the current `LATEST` version, i.e. `4.0`.

This change fixes those tests. It was not clear if those tests required
a specific version but just in case they did they were fixed by adding
`format = FileFormat.V4`.

Bug: 295577788
Test: ./gradlew
Change-Id: I98f0acb4ed2a5aa682d570d9d37da522bd591666