Refactor ChromeSliceDetailsTab components for  Scroll Jank plugin.
The EventLatency track used to reuse the ChromeSliceDetailsPanel,
but we want to include custom linking - as such, this patch
refactors out the common ChromeSliceDetails component generation
to a set of utils functions and reuses them in a new details panel
for the event latency track.

Because the track now has its own details panel, linking/selection
can be implemented in a generic way (can be reused to link to other
patches). As such, replace the EventLatencySlice with a more
generic ScrollJankSlice.

Add a link to the original slice in the renderer process.

https://screenshot.googleplex.com/Bh7xQm2ZyBKe3oV

Bug: b/278844325
Change-Id: I91bfd702abf0ada2fe7b79371db94dcb8c1814f5