Make VERSION and HASH inline static member vars
CPP backend needed both to be static inline, and the ndk backend
needed the version number to be inline since HASH already was.

These variables are changed to static to fit with other backend
behavior and allow processes to easily get the version of the local
interface.
These variables are inline so they don't need to be defined in the
.cpp files to satisfy the linker.

Test: m
Bug: 295582637
Change-Id: Ia34a8020104ef6fdff25d39a0725c7c7984fa0c8