Move TextTestUtils into metalava-model-text test fixtures
This makes uses of the `java-test-fixtures` plugin to allow projects to
share testing related code to each other. It moves the
`TextTestUtils.kt` file into `metalava-model-test`'s test fixtures
directory. In the process it removes the unnecessary `stripComments`
call.

Text fixtures cannot be published to Maven and gradle would report a
warning every time it was used. We do not actually want them published
to Maven so  this suppresses them.

Bug: 294556178
Test: ./gradlew
Change-Id: I36717d7e64c987f8824c5e7e0f9e8af2bb31e06a