Default to use 2.0 if no signature header is provided
Previously, a lot of tests relied on v1 having no signature header.
However, as v1 is being removed they would start failing if left alone
so this change ensures that a 2.0 signature header is added if no
header is present. Assing a line at the beginning of the files caused
quite a lot of churn in line numbers used in error reports.

Bug: 296047828
Test: ./gradlew
Change-Id: I301fbfb43bb191e3f45104f8a845da8bd38c8d7d