Allow format specifier to be used in signature files
The existing versions do not actually support a format specifier in the
file, at least not in the sense that a previous version of metalava
which supported that version would fail to parse a file of that version
with a format specifier. However, it will be useful for migration to
allow a specifier to be temporarily added to the signature file (maybe
with a warning that it is being migrated).

The idea is that if metalava currently does `X` by default but we want
it to do `Y` then we would:
* Add a new `X-or-Y` property to the format specifier, defaulting to
  `X`.
* Update all the signature files to specify that default, e.g.
  `X-or=Y=X`
* Then switch the default to `X-or-Y=Y` in metalava. That should have
  no impact on the files as they specify `X-or-Y=X`.
* Migrate each file separately to `Y`, which would remove the property
  and just use the default.
* Remove the `X-or-Y` property.

Bug: 295577788
Test: ./gradlew
Change-Id: I37048032279aa544c4afbe1f8eda23e14aa23afa