Detect when --api-overloaded-method-order is unspecified
Previously, if the option was unspecified then it would just use the
default of `SIGNATURE` which meant it was impossible to determine if
the option had been set. That is important because eventually it will
be necessary to allow other sources to provide the
`overloaded-method-order` property while still allowing the command
line option to override it.

This change adds support to `enumOption` to specify a default that is
hidden from the help but still allows the help to pretend that there
is a default for the sake of documentation.

Bug: 295577788
Test: ./gradlew
Change-Id: Ia72855c0e214396c3ffd106b53e1d0be9e85c269