Stop calling Driver.run to test CliktCommands
Previously, the `BaseCommandTest` just called `Driver.run()` and let it
do all the work in creating the `MetalavaCommand` and all the
subcommands but that meant it was doing more work than necessary. So,
this change makes `BaseCommandTest` create the `MetalavaCommand` and
add the subcommand that is being tested.

As the subcommands all define their options using Clikt they do not
need to set `treatUnknownOptionsAsArgs = true` so this modifies
`MetalavaCommand` so it only passes true for that iff it has a
non-null `nonCliktUsageProvider` as that indicates that it also has its
own non-Clikt option processor.

Bug: 295829461
Test: ./gradlew
Change-Id: I7feea4bef4b5d207bb8f8d4de1b81eac063bd947