Remove metalava sandboxing
It was added in https://r.android.com/1298054 to simplify migration of
metalava from having free access to all files to running inside a
sandbox. It could report or fail on unexpected file reads, where an
unexpected read is one that accessed a file which was not explicitly
listed on the command line.

However, now that it is running inside a sandbox in the Android build,
AndroidX never used it, and Bazel already supports sandboxing there is
no need for Metalava to continue to provide its own which is, always
going to be less protective than a purpose built generic one.

Also, the Metalava sandbox is built on Java's `SecurityManager` which
is scheduled for removal so it already has a limited life span as it
is.

Bug: 296353397
Test: ./gradlew

However, now that it is running inside a sandbox in the Android build,
AndroidX never used it, and Bazel already supports sandboxing there is
no need for Metalava to continue to provide its own which is

Change-Id: Ia3482a2233696ef3566227e217d06279e023eb25