Move post command cleanup into DriverCommand
The moved code only has any effect when `options` are parsed and errors
are reported to one of the `DefaultReporter`s in `allReporters` which
only happens in the `DriverCommand`.

Ensuring that the cleanup code executes at exactly the same time as it
did before, i.e. after any `MetalavaCliException`s were handled but
before exiting required the addition of a special mechanism to support
post command cleanup.

Bug: 296362159
Test: ./gradlew
Change-Id: Ic29aab705912e85b833412757ef6036cbbe6ac93