Manual merge ANGLE from ffe81dd3cd63 to f9e3f6743063 (20 revisions)
https://chromium.googlesource.com/angle/angle.git/+log/ffe81dd3cd63..f9e3f6743063

Manual merge was needed because:
https://chromium.googlesource.com/angle/angle.git/+/0f4317d01cdaed157ab999934839c554bfa6cefd
added .gitmodules file in third_party/vulkan-deps,
so it couldn't be submitted.

The next commit
https://chromium.googlesource.com/angle/angle.git/+/d52046c499c0238a15373b7b5aa550b07be3a50b
couldn't be merged automatically, because it modified
third_party/jdk/README.chromium, which is deleted in AOSP.

The merge is up to
https://chromium.googlesource.com/angle/angle.git/+/424f43e40e14f7f7362ef06948b0bb0cc2917c74
which causes hangs in AOSP tests, e.g. dEQP.

Test: Presubmit checks will test this change.
Change-Id: Id96babf56fc630d4957a6e61dee425fcace86763