Separate some Clikt patches from the main CLI code
It has been necessary to workaround some limitations in the Clikt
library (mainly due to `metalava` doing some weird stuff). Previously,
those workarounds were mixed in with the rest of the code. This change
starts the process of separating them out which should make them easier
to address when upgrading Clikt.

It moves the `addHelpParams()` and `parameterHelp()` methods from
`MetalavaCommand` and `MetalavaSubCommand` to `CliktPatches` and makes
them work on all `CliktCommand`s not just specific subclasses. That
requires adding a new `shortHelp()` method.

Bug: 295829461
Test: ./gradlew
Change-Id: If850f60edfda572c242df38cff5287a77ed770bf