Autoupdate of 2 submodules
*external/python/cpython3 updated to 02afc01277f68e081dad208f2d660fc74d67be88
Simplify the launcher path detection (prep for 3.11)

As of 3.10, we've had the ability to modify the `lib` name in the
default lookup path, so change that to `internal` to match our existing
layout. Then move the stdlib to the versioned directory
`internal/python3.10` via pkg_path. This removes the need for us to add
custom entries to the path and entirely use the default set.

Remove the src or prebuilt stdlib logic from Lib/Android.bp, it's being
moved into soong so that we can have different `pkg_path` settings
between the two.

Bug: 278602456
Test: treehugger
Test: build/soong/python/tests/runtest.sh
Change-Id: Ice483a8102d02bfbfbca2c66c7a1371b300dd4de

*prebuilts/build-tools updated to 9ca5172c8800328101c15fd6d9d135053c44c53e
Merge "Define a python_library for the prebuilt stdlib" into main