Change how hidden package filtering works
Instead of waiting to filter hidden packages until after java and kotlin DPackages are merged, this now removes the java packages before the merge, and then after the merge removes any remaining DPackages with the same name as an already removed package. This is because the java version of the package could be filtered out before the merge by a different filter, so after the merge there would be no way to know the kotlin version package should be hidden, since it was never merged with the java package that had that info.

Bug: 296287991
Test: I couldn't get the bug to reproduce in an integration test--added one that theoretically should reproduce it but didn't. I ran `./gradlew shadowJar` and use that jar to run `docs-tip-of-tree` for AndroidX, and confirmed that I was able to remove `@suppress` from health classes and they did not appear in the docs.
Change-Id: Ib3a20e6e8c73f87ec3ebe6015e03412d4e5a83f1