Merge hide property into Showability
Fundamentally, the hide and show properties are mutually exclusive both
at the individual annotation level (a single annotation can only be a
hide or a show annotation not both), and at the `Item` level, an `Item`
cannot be shown and hidden from the same API surface.

Previously, they were separate properties (`hide` and `showability`).
This change removes the `hide` property and adds a new
`Showability.HIDE` instead. When an `Item` has both show and hide
annotations (there are tests that cover that) then it will resolve in
favor of the show which is what the previous behavior did but was not
very clear.

Bug: 292484829
Test: ./gradlew
Change-Id: I0ee2ede32bb27425adcfd78f1a56f5a43b28469b