Fix test case when no server address is available from system.
When this test case is run on devices which do not support Remote Key
Provisioning, it fails as there is no URL property configured. This
causes a MalformedURLException when the tests try to generate the URL
for fetching information from server. With this change, we will mock
the server info so that we can safely bypass that exception and test
the actual failure.

Bug: 296809789
Test: atest RkpdAppUnitTests
Change-Id: I7a91523c0ea7447f974e40c9636e36461dcc449e