Update bug component from Android Systems to OTA client
It's been a while since the Android Systems bug component does not
access the creation of new bugs for triage, instead, new bugs must
be created in the subcomponents of the team.
This change updates the bug component originally associated to this
subsystem from the Android Systems' 30545 to OTA client's 1014951.

Test: none
Bug: 270571229
Change-Id: I10ea8fba3b7683cf6259394f2163b68bfc96dfb5
Signed-off-by: Alessio Balsini <balsini@google.com>