tp: allow conversion from int to double in runtime tables
It's not uncommon for tables to queries which populate tables to
start off containing ints and then switch to doubles or vice versa.
Account for this case by adding code which converts ints to doubles
if there is a mixed column containing both. Still ensure that all ints
are abs(i) <= 2^53 as this is the maximum number densely representible
by doubles

Bug: 296599532
Change-Id: I816202c464a0b99cd02b8b5ddb3c4b372fabed10