Enable intrinsic optimization for polymorphic intrinsics
Polymorphic invokes are expensive and some of the methods in Atomic*
classes uses polymorphic methods. We use intrinsics to generate
efficient code for them. Intrinsic optimization was disabled in
debuggable runtimes to be able to support breakpoint in intrinsic
functions. It might be less useful to break in such methods so we want
to enable intrinsic optimization for polymorphic invokes which are
performance sensitive.

Bug: 296298460
Test: art/test.py
Change-Id: I575695d82e8bc7d703cfbf5ff22ea7d5a35f6937