Adjust symlink to point to updated Linux prebuilt runtimes.
We recently cleaned up the symlink paths by getting rid of the "17.0.3"
path in favor of just "17". Unfortunately, we neglected to update this
specific Darwin-only symlink to the Linux prebuilt runtimes. This only
breaks Darwin builds that are relying on the compiler-known paths, and
thus we didn't notice the missing update in our host Darwin builds for
everything else (other than Rust).

Test: prebuilts/python/darwin-x86/bin/python3 toolchain/android_rust/build.py --llvm-linkage shared --upstream-test --ndk toolchain/prebuilts/ndk/r25/
Change-Id: I1d326d21cc139ae5eca729c20df163f194478997