Update bug component from Android Systems to Kernel Tests
It's been a while since the Android Systems bug component does not access the
creation of new bugs for triage, instead, new bugs must be created in the
subcomponents of the team.
This change updates the bug component originally associated to this subsystem
from the Android Systems' 30545 to Kernel Tests's 391836.

Test: none
Bug: 270571229
Change-Id: Iace5c1adfda4ba0ec4bc3bff84fc624f8325d415
Signed-off-by: Alessio Balsini <balsini@google.com>