PvmfwImgTest: Use actually invalid config data version
Since aosp/2714274, configuration data blobs of a version v1.x more
recent than the latest version known to pvmfw are being accepted to
provide forward compatibility with the system bootloader. As a result,
testInvalidConfigVersion_doesNotBoot is now failing because the version
it uses (v1.100) has become valid so that the pVM boots successfully.

Instead, use v2.0, denoting a breaking change in the format, which must
therefore be rejected by pvmfw (until we introduce said version).

Bug: 296207806
Bug: 297342022
Test: TH
Change-Id: I6725f7a4e081c598b8b5cfd78d561f3b0e6ce0f6