riscv: Set loadaddr first in board_late_init()
RISC-V u-boot currently doesn't provide a device-tree for "kernel-start"
info. This causes board_late_init() to exit early before setting the
loadaddr env variable. This causes EFI bootflow to fail as it expects
loadaddr to be set. For now, we move the setting of loadaddr to earlier
in board_late_init() to workaround the problem, similar to qemu_arm.c.

Test: lunch aosp_cf_riscv64_phone-userdebug. launch_cvd -bootloader=<new
bootloader> -android_efi_loader=<grub.efi>

Bug: 294059825
Change-Id: I8b6ede536fdd8c24a5e43677819a9d98b9836d46