Add configuration for which annotations hide documentables
This will be a breaking change for any users relying on the current list of annotations that hide things (like gmscore), so I'll send an email out when we next do a version bump.

We can add the new config value to AndroidX now (aosp/2726214) so it will be ready for the version update.

Test: integration test with custom hide annotation, all prebuilts tests using @RestrictTo still pass
Bug: 275544802
Change-Id: I7e285fc0cdfbc52cceb296df330c9726ceaeb64c