Autoupdate of 2 submodules
*external/icing updated to ec9c4f473d9b5b6d316405f5057eeeddbaa27ff5
Move icing to be a simple java project

Since icing is only bundled inside of appsearch:appsearch-local-storage
there is no need to use an android project as at the end of the day we
end up only including the jar of icing anyway. Moving to a java project
simplifies is as now we no longer need to add special export
configurations.

We are also moving the dropping of *.proto files to the projects
that bundle the libraries.

This is a step towards making it easier to resolve what artifact
coordinates were actually bundled inside.

Test: ./gradlew appsearch:appsearch-local-storage:publish
      comparing out/androidx/build/support_repo/androidx/appsearch/appsearch-local-storage/1.1.0-alpha04/appsearch-local-storage-1.1.0-alpha04.aar
      before and after the change, libs/repackaged.jar is identical
Bug: 296864329
Change-Id: Ibba769f71ecdb58cb20df0ea0c37364cf579eddf

*frameworks/support updated to f50b1d036f2fcfd30406ac281f562b3d9c6124e5
Merge "Add support for dropping java resources in BundleInsideHelper" into androidx-main