Extend the time to download MTS zip file.
Context:

We set up a partner lab for OEM partner to run the mainline MTS testing via ATS,  which will download the andorid-mts.zip file from the GCS.  The MTS file is ~5.3G , and the bandwidth between the partner ATS lab and GCS is only 25MB/s to 30MB/s. which takes sometimes slightly longer than 30 mins to download the files. So the current download timeout is not enough and will block the testing.

Change-Id: Id620ee958bebe6b5b6715f537febd7aa1e295e84