Make overloaded-method-order override --api-overloaded-method-order
Previously, if specified, `--api-overloaded-method-order` would always
override the `overloaded-method-order` property set in the format
specifier. This change switches the precedence so that the
`--api-overloaded-method-order` is only applied if
`overloaded-method-order` has not been explicitly  specified.

The reason for this change is it will allow a signature file to specify
that it uses `signature` order while other files, which do not specify
a particular order to continue to use source order that is specified
by `--api-overloaded-method-order=source`.

Bug: 295577788
Test: ./gradlew
Change-Id: Ie609f0ccac3993742ecea86c823a189511796e32