Push psi specific code in loadFromJarFile(...) into PsiSourceParser
Previously, `loadFromJarFile(...)` had some Psi specific code to create
the `PsiBasedCodebase`. This pushes that code down into
`PsiSourceParser.loadFromJar(...)`.

That allowed the `annotationManager` parameter to be removed from
`loadFromJarFile(...)` as the `PsiSourceParser` already has its own
`annotationManager` property. However, that did require some slight
changes in `ConvertJarsToSignatureFiles` as it previously passed
`noOpAnnotationManager` (the default) into `PsiSourceParser(...)`
which was never used and then passed `DefaultAnnotationManager()` into
`loadFromJarFile(...)` which was used. As the former is now used and
the latter is not it was necessary to pass `DefaultAnnotationManager()`
into `PsiSourceParser(...)`.

Bug: 298092713
Test: ./gradlew
Change-Id: I35c24dffe960a3968025ac4b95ea2c1a0e99e897