Move creation of PsiSourceParser into PsiEnvironmentManager
Previously, the `PsiSourceParser` was created directly by the
`processFlags(PsiEnvironmentManager)` method. This change moves that
into a `PsiEnvironmentManager.createSourceParser(...)` method that
implements the new `EnvironmentManager` interface's method.

One slight complication is that the `LanguageLevel` and
`LanguageVersionSettings` that were passed into `PsiSourceParser(...)`
are Psi specific. So, this change passes them in as strings and then
the `PsiEnvironmentManager` converts them to the types expected by
`PsiSourceParser(...)`. It also corrects the options help which
incorrectly reports that kotlin defaults to `1.9` when in fact it is
`1.8`.

Bug: 298092713
Test: ./gradlew
Change-Id: I9c17bb83f6362d614a9c1fada604cef976b60c7e