Fix android-jars-to-signatures when loading android.txt file
Previously, when `android-jars-to-signatures` found an `android.txt`
file corresponding to an `android.jar` file and tried to load it it
would throw an exception because it was trying to use an initialized
Clikt option when accessing `options.annotationManager`, which is the
default value of the `SignatureFileLoader.load(...)`'s
`annotationManager` parameter.

This change updates the test to reproduce the problem and then fixes
the problem by explicitly passing an `annotationManager` thereby
avoiding using the default value.

Bug: 298200709
Test: ./gradlew
Change-Id: I3b40f4da42fe9b27942c05775de9f9e0c3df6927