Remove tick() and Reporter.showProgressTick()
This was globally accessible and so tightly coupled any code that used
it. The alternative was to just stop it being accessed globally and
instead pass an object around (like `Reporter.showProgressTick()` but
as this is not relied upon by existing users that would be pointless.
Also, `showProgressTick()` does not really fit in with `Reporter`.

Bug: 295013652
Test: ./gradlew
Change-Id: Ie8e8873e264bcd9ce0470121feb1612897a3f965