Remove package ProgressTracker instance and progress(String) method
Previously, any code that wanted to call `ProgressTracker.progress()`
did via the package `progress(String)` method. This change removes that
and instead passes the `ProgressTracker` instance around to where it
is needed.

This also removes any additional checks of `options.verbose` that wrap
the call to `progress(String)` as they are unnecessary because the
method already checks `options.verbose`.

Bug: 295013652
Test: ./gradlew
Change-Id: Ie4e8ebb611da300a87b632906e3dba677ab9313d