Stop SignatureWriter depending on `options`
Previously, `SignatureWriter` would access `options` indirectly via the
`ApiVisitor`, either through its default parameter value of
`showUnannotated` or because it uses `options.stubPackages` in the body
of one of its methods.

This change allows them to be specified as parameters to `ApiVisitor`
and/or `SignatureWriter` so that they can be provided by the callers.

Bug: 286023667
Test: ./gradlew
Change-Id: I99b82f611905fe8e11cf7e610e2055b2f4dac4c6