Allow tests to specify classPath to testCodebase
The `createTestCodebase` method was adding kotlin stdlibs and an
android jar to the `classPath` passed to `parseSources(...)` but the
kotlin stdlibs were not used at all and the android jar was only used
by one test. So, this change removes that from `createTestCodebase`
and allows the tests to supply their own `classPath`.

Bug: 294556178
Test: ./gradlew
Change-Id: I741aa5376ff6c6a902b2cb29f4195f02e6145ecf