Add an option to allow test mapping suite to ignore certain keywrords
The rational is that, we plan to split tests in presubmit group into
aosp CF and non-aosp CF, and that only applies to main branch. For other
branches, all tests will still run on a single CF target, which means it
needs to run all tests in presubmit group regardless the setting of
keywords.

Bug: 298228905
Test: m tradefed-all, unittest

Change-Id: Iaf7acb95832d36867b53a5a847a17b8edc9b27c5