gfxstream: nuke uses of puid for VK tracing
This is more less an revert of "[8/n] vulkan: add trace point
in vkQueueSignalReleaseImageANDROID".  The overall rationale
is to remove the need for the auxiliary renderControl stream
for atleast for VK Cuttlefish/Trout/Fuchsia guests.  In those
instances, the virtio-gpu ctx id is the PUID and the extra
side channel isn't needed.

CAPSET_GFXSTREAM_VULKAN will provide versioning in subsequent
patches.

Also, if we want generic guest/host traces, we'll probably want
to call this in geneirc ReleaseImage API, not just QSRI_ANDROID.
Though AFAIK nobody really uses combined guest/host tracing
at the moment and could be an artifact.

BUG=297245651
TEST=launch_cvd

Change-Id: I1d3c096569ebdc3f6eec343efbd7cccff1c4849e