Allow 5.0 to be customized without migrating
Previously, all the versions (apart from `5.0`) did not support
customizing via properties embedded in the signature. So, it was only
possible for them to do that temporarily while migrating, just in case
they were consumed by a slightly older version of metalava that did
not understand the properties. However, `5.0` supports properties
properly so there is no reason to limit the customizing just for
migrating. So, this change relaxes that restriction for `5.0`.

Bug: 295577788
Test: ./gradlew
Change-Id: Ibfa80d94d4b5afa5273ed7f9633172ad88cc9130