Add header and specifier tests for FileFormat
Previously, the parsing of header and generation of headers were tested
separately in `FileFormat` and specifier parsing was not tested at all,
that was left up to `SignatureFormatOptionsTest`.

This change adds tests (and combines a couple of existing tests) that
test all aspects of reading and writing both headers and specifiers.

Bug: 295577788
Test: ./gradlew
Change-Id: If7cda5e7a0b56d013f264de22f11df7df77053e8