Add FileFormat.surface property
The `merge-signatures` command is only intended for merging signature
files of the same API surface from different libraries but
unfortunately, had no way of verifying that was the case. This change
addresses that by adding the `surface` property which can be embedded
in a signature file (for versions 5.0 and above).

The `surface` property is simply a label for the API surface that the
`merge-signatures` command can check. At the moment it is optional but
the intent is that it will become mandatory in future.

A follow up change will update `merge-signatures` to make use of this
new property.

Bug: 300043632
Test: ./gradlew
Change-Id: I2ecde38d69292ef585a87786ae5d3fabeaf099ba