Remove Context parameter from copy of CliktCommand.allHelpParams
The private `CliktCommand.allHelpParams` method was copied (in so far
as it could be) for use by metalava to support combining the help for
the default command, i.e. `DriverCommand` with the general help for
`MetalavaCommand`.

The copy had a number of limitations over/differences to the original:
1. It did not support hidden commands
2. It took an extra `Context` parameter.

This change removes the extra `Context` parameter so that a follow up
change can invoke the private `allHelpParams` method to support hidden
commands properly.

It required two extra fixes when generating the combined help for
`MetalavaCommand` + `defaultCommand`. They were not required before
because the combined help was generated using `MetalavaCommand`'s
`currentContext` only but now it is generated using both it and the
`defaultCommand`'s `currentContext`. The change are:
1. The `defaultCommand` needed to have its `currentContext` initialized
   explicitly.
2. The `DriverCommand` needed to set its `localization` explicitly.

Bug: 286023667
Test: ./gradlew
Change-Id: Ib889e26ad69cc0892d766bfcbfb5ca41da69bf8e