Add baseline file for metalava-model-test-suite failures
Previously, the `@IgnoreForProvider` annotation was added to allow
tests to be disabled for a specific provider to allow the provider to
be implemented incrementally. Unfortunately, there were some problems
with that approach:
1. It required changing the tests twice, once to add (or change the
   providers listed) to disable the test and again when re-enabling it.
2. It becomes a source of conflict when adding multiple providers or
   even just adding new tests for existing providers.

This change adds support for using a baseline file per provider to
ignore any tests that are expected to fail. That means it is not longer
necessary to modify the tests eliminating the issues with the previous
approach. It also has a number of other benefits:
1. It makes it easier to identify the list of remaining tests that need
   fixing.
2. Updating the baseline tests to remove all tests that are now passing
   is done automatically.

The basic idea is that you run the tests as normal and then run the
following afterwards:
```
./gradlew updateModelTestSuiteBaseline
```
That will update the baseline by:
* Adding tests that are failing.
* Removing tests that are working.

When adding a new test that breaks on some providers simply run:
```
./gradlew test --continue
./gradlew updateModelTestSuiteBaseline
```

That will update the baseline for all affected providers.

When extending a provider implementation to fix a test then once the
test is passing run the above again and all tests that are now fixed will
be removed from the provider's baseline.

Bug: 294556178
Test: ./gradlew
Change-Id: Ifb89048b5303f556ddbef28a1a406f5f9b02f07f