Fix slice "view culling" bug in process_scheduling tracks.
The pointer hit detection algorithm expects slices need to be in
[cpu, time] order, but the renderer's view culling algorithm expects
slices to be in [time] order as it performs a binary search based on
slice start time to find the last slice to render.

Since aosp/2585517, I changed the order from [time, cpu] to [cpu, time]
which fixed the mouse hit detection but broke the rendering, resulting
in a lot of slices not being rendered depending on the position of the
visible window.

This CL fixes the view culling algorithm by removing the binary search
in favour of attempting to render all slices but aborting early if a
slice is outside the visible window. The cost of an aborted render is
negligible compared to the cost of actually rendering the slice, so the
performance overhead should be negligible.

Bug: 298945371
Change-Id: I471dd59f9251767ef7b188cbc547e0d4dbf130fe