Autoupdate of 2 submodules
*system/libvintf updated to fa5d486de907e9b2bae40ad68893c72234df5733
Remove the check against updatable-via-apex in FCM

The attribute updatable-via-apex was to control which HALs can be put in
Vendor APEX. However, as we're migrating every HAL in AOSP into Vendor
APEX, the attribute has less meaning.

Hence removing the check against the attribute in FCM, which means every
HAL can be put in Vendor APEX.

Bug: 298873490
Test: libvintf_test vintf_object_test
Change-Id: Ie80b672dfd048fd39682bbcd719ebede0b385edc

*test/vts-testcase/hal updated to 6d77507eb6aad0600308e6f95c5393d1c53d8d64
Revert "Add VTS test to check updatable-via-apex for updatable HALs"

This reverts commit 655734e10f00fd8d88ba347ef02394ba2059f184.

The attribute updatable-via-apex was to control which HALs can be put in
Vendor APEX. However, as we're migrating every HAL in AOSP into Vendor
APEX, the attribute has less meaning.

Hence removing the check against the attribute in FCM, which means every
HAL can be put in Vendor APEX.

Bug: 298873490
Test: DeviceManifestTest
Change-Id: Iaab8123c253628f64ff68ea0453cdb16141d7836