Set accept_ra_min_lft systctl when starting IPv6 stack.
The sysctl accept_ra_min_lft is used to tell the kernel to ignore
any individual RA section with lifetime below this value. Set it
to 3 mins initially when starting IPv6 stack if it's supported by
kernel. The value is configurable via flag push. This is helpful
to prevent cpu from waking up frequently every time when a RA is
received (actually APF filters most of the duplicated RAs and
ignore any RA section with low unreasonable lifetime with the aim
of this systcl).

Bug: 291182564
Test: atest NetworkStackIntegrationTests
Change-Id: I6fd5274e45a14fdf7952c355d19c57d75e541b62