eliminate all support for netd vs libnetdclient desync
These were recently added to support libnetdclient and netd
to be built at different points in time, but this is not
something we want to support.

The sysprop is a mechanism for OEMs to configure which syscalls
are shimmed, but that no longer affects the operation of
firewall mark server itself, since the size of the inbound
request no longer changes depending on it.

As such the server no longer checks the ro.vendor.redirect_socket_calls sysprop: it always expects the destination address for ON_CONNECT, ON_CONNECT_COMPLETE, ON_SENDMSG, ON_SENDMMSG, ON_SENDTO.

The client is simply expected to not send any of the latter 3
if the sysprop is false.

In practice - when compared to U - this is really only a change
in behaviour for ON_CONNECT - previously it only had a dstaddr
specified if the sysprop was true, now that's unconditional.

Note: ON_CONNECT_COMPLETE has always required dst addr.

Note2: ON_SEND{MSG,MMSG,TO} have always passed in dst addr,
but would simply not happen if the sysprop was false.
The server would expect them to have dst addr if the sysprop
was true.  In effect there was a harmless disconnect here,
when the sysprop was false in what the server expected
and what the client would do *if* it was changed to actually
issue these...

Test: TreeHugger
Bug: 150126287
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I4dd4440825e6ef3243663970721d829d48028e0d