Relax the rules about what constitutes a valid signature file
Some previous changes inadvertently tightened the rules as to what
constitutes a valid signature file. It only affected signature files
which were manually edited, not those created by metalava.

The failing signature files had text after the `// Signature format`
line that was either not `package ...` and not the end of the file.

This change adds some tests to cover those examples which reproduced
the problem and updates the code to fix them.

Bug: 295577788
Test: ./gradlew
Change-Id: I11f7495321c74212c63909a88f59141399346a96