Ensure options is set before calling Driver.run
Change https://r.android.com/2736560 allowed code to disallow access
to the `options` property to ensure that it was not being used
unexpectedly. Unfortunately, that broke some tests that did not
explicitly initialize the `options` before using it.

e.g. The `DriverTest.check(...)` method calls `Driver.run(...)`.
Although, that does set `options` in `DriverCommand.run()` it actually
accesses it prior to that. That access needs cleaning up but in the
meantime this sets the `options` before calling `Driver.run(...)`.

As the test failure was very intermittent I tested this by setting
`OptionsDelegate.possiblyNullOptions` to `null` by default and running
`StubsClassTest`. All the tests failed. They passed with this change.

Bug: 286023667
Test: ./gradlew
Change-Id: I446467d70492584ac273a212689af3acc4b18cea