Move `Options.getUsage` into `OptionsHelp` object
The help methods in `Options` were instance methods but that meant it
was necessary to access them through the `options` property.
Fortunately, they do not actually use any instance state so while they
could have been moved into a companion object that would have changed
the indentation and caused a huge delta. So, instead they are moved
into a new standalone `OptionsHelp` opject which thanks to them being
the last methods in the `Options` class produces a very small delta.

Bug: 286023667
Test: ./gradlew
Change-Id: I085e0d1205df9f354d92638d67f0d369dc610ea8