Move `verbosity` into a new `EarlyOptions` group
The `EarlyOptions` is processed before most of metalava runs so that
metalava can enable verbose logging as early as possible. That
eliminates the need to access the global `options` property before
running the main commands.

One minor side effect is it changed the order of properties in the
command line help because properties from extended classes are
initialized (and so registered) before those in the extending class and
`verbosity` was not the first property in `CommonOptions`.

Bug: 286023667
Test: ./gradlew
Change-Id: I85f7dab00a4c19ddfeab4dcc5d4661c87bd57699