Add ApiPredicate.Config to pass options to ApiPredicate
At the moment it just contains two properties that were not passed in
to the constructor at all and which were just initialized in the
default value. They are now initialized in `options.apiPredicateConfig`
and that is used in the default parameter value for `config`. It still
needs to be inside `try...catch...` block to prevent test failures.

Bug: 299280167
Test: ./gradlew
Change-Id: I532249eab21bf180a7b073162cbc43a915d66c7c