Reduce the time needed for SensorManager EventQueueStress
Only enable/disable the first eligible sensor for each event queue. As
long as one is enabled succesfully, the object will be set up in the
framework and will need to be destructed before it hits the file limit.
This was previously done for the AIDL version of this test, and now this
is for the HIDL version of the test.

Test: m
Bug: 299171035
Change-Id: Icf18450812e1aa3ae26217fec5b171841083d12b