Move ignoreShown parameter into ApiPredicate.Config
Although `ignoreShown` can come from `options.showUnannotated` that
can be overridden with a hard-coded value by the code creating the
`ApiPredicate`. This uses the data class `copy(...)` method to change
the `ignoreShown` in the `Config` object where necessary.

In a number of places it was necessary to convert the implicit
dependency on `options` into an explicit dependency. Although, that is
increasing the number of references to `options` (which we're trying to
reduce) it is still a step in the right direction as it is still
pushing those dependencies down the call stack and out of common code
shared by lots of commands into `DriverCommand` specific code.

In a few places it was challenging to find a place to add the
`@Suppress("DEPRECATED")` annotion so a little refactoring was needed.

Bug: 299280167
Test: ./gradlew
Change-Id: I687b4582141d5fbc162f950cee1fb3fa658dfc8d