Add ApiPredicate.Config to ApiVisitor.Config
This adds `ApiPredicate.Config` to `ApiVisitor.Config` and makes sure
that `options.apiVisitorConfig` uses `options.apiPredicateConfig`.

As `options.apiPredicateConfig` depends on
`SignatureFormatOptions.fileFormat` which depends on Clikt options it
will fail if the `options` class has not been parsed by Clikt which
only happens in `DriverCommand`. So, that breaks the tests for a few
commands. This change prevents that by making sure that an empty
`ApiVisitor.Config` is passed through where necessary.

Bug: 299280167
Test: ./gradlew
Change-Id: Ib05ccbf72783c863c71967b489362564772c0fc3