Add FileFormat.formatDefaults to support --api-overloaded-source-order
Previously, the `--api-overloaded-source-order` was added to the format
by `applyOptionalCommandLineSuppliedOverrides` which would only use its
value if the format did not explicitly specify it. The issue with that
is it made it lost the information on whether the property was
explicitly specified.

This change adds an optional `formatDefaults` property that contains
defaaults that are added to the command line. At the moment it is only
used for `--api-overloaded-source-order` but a following change will
add support for controlling the defaults of other options.

Bug: 300052204
Test: ./gradlew
Change-Id: I1eccee20a33d8b64a68c25759fb9e0fa0f95bc94