Add FileFormat.name property
The `merge-signatures` command is only intended for merging signature
files of the same API surface from different libraries but
unfortunately, had no way of verifying that was the case. This change
addresses that by adding the `name` property which can be embedded in a
signature file (for versions 5.0 and above).

The `name` property is simply a label for the API that the
`merge-signatures` command can check. At the moment it is optional but
the intent is that it will become mandatory in future.

A follow up change will update `merge-signatures` to make use of this
new property.

Bug: 300043632
Test: ./gradlew
Change-Id: I11bd3c76ae1f4b39c4fc7615434d82f3644dc3c3